Vinculando la base de datos con el Visual y corrigiendo errores

+++ database/models/Hotel.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false,
         },
         nights: {
-            type: DataTypes.Decimal,
+            type: DataTypes.INTEGER,
             allowNull: false
         }
     };
